refactor(15311): refactor the series computation out of the individual charts
